requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

# pbr should be first
pbr>=0.6,!=0.7,<1.0

argparse
iso8601>=0.1.9
oslo.config>=1.6.0  # Apache-2.0
oslo.serialization>=1.2.0               # Apache-2.0
oslo.utils>=1.2.0                       # Apache-2.0
PrettyTable>=0.7,<0.8
pycrypto>=2.6
requests>=2.2.0,!=2.4.0
simplejson>=2.2.0
Babel>=1.3
six>=1.9.0
python-keystoneclient>=1.1.0
